refactor: extract to transform
